requirements.txt
python-binance
pandas
python-dotenv
ta
numpy
schedule
requests
